requirements.txt
certifi==2020.6.20

pillow~=8.1.2
numpy~=1.23.0
pygame~=2.1.2
pexpect~=4.8.0
opencv-python~=4.6.0.66
pypylon~=1.8.0
picamera~=1.13